refactor cli rpc directory
