+++ wi/wi-extension/src/cloud/cmds/sign-in-with-code-cmd.ts
@@ -20,7 +20,7 @@ import { WICommandIds, type ICmdParamsBase } from "@wso2/wso2-platform-core";
 import { type ExtensionContext, commands, window } from "vscode";
 import * as vscode from "vscode";
 import { ResponseError } from "vscode-jsonrpc";
-import { ErrorCode } from "../../choreo-rpc/constants";
+import { ErrorCode } from "../choreo-cli-rpc/error-utils";
 import { ext } from '../../extensionVariables';
 import { isRpcActive, setExtensionName } from "./cmd-utils";
