+++ wi/wi-extension/src/extension.ts
@@ -17,20 +17,9 @@
  */
 
 import * as vscode from "vscode";
-import type { ConfigurationChangeEvent } from "vscode";
-import { authentication, commands, window, workspace } from "vscode";
-import { WSO2AuthenticationProvider, WSO2_AUTH_PROVIDER_ID } from "./cloud/auth/wso2-auth-provider";
-import { ChoreoRPCClient } from "./choreo-rpc";
-import { installRPCServer } from "./choreo-rpc/activate";
-import { getCliVersion } from "./choreo-rpc/cli-install";
-import { activateCmds } from "./cloud/cmds";
+import { activateCloudFunctionality } from "./cloud/activate";
 import { ext } from "./extensionVariables";
 import { StateMachine } from "./stateMachine";
-import { contextStore } from "./cloud/stores/context-store";
-import { dataCacheStore } from "./cloud/stores/data-cache-store";
-import { locationStore } from "./cloud/stores/location-store";
-import { activateURIHandlers } from "./uri-handlers";
-import { getExtVersion } from "./utils/commonUtils";
 
 /**
  * Activate the extension
@@ -58,95 +47,6 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
 			`Failed to activate WSO2 Integrator Extension: ${error instanceof Error ? error.message : String(error)}`,
 		);
 	}
-}
-
-/**
- * Boot the cloud-connected functionality — mirrors the platform extension's activate():
- * cloudEnv → store rehydration → auth provider → RPC client → initAuth → config → pre-init handler
- */
-async function activateCloudFunctionality(context: vscode.ExtensionContext): Promise<void> {
-	// 1. Resolve cloud environment
-	ext.cloudEnv =
-		process.env.CHOREO_ENV ||
-		process.env.CLOUD_ENV ||
-		workspace.getConfiguration().get<string>("WSO2.WSO2-Platform.Advanced.ChoreoEnvironment") ||
-		"prod";
-
-	// 2. Log versions
-	ext.log(`Extension version: ${getExtVersion(context)}`);
-	ext.log(`CLI version: ${getCliVersion()}`);
-
-	// 3. Rehydrate persistent stores
-	await contextStore.persist.rehydrate();
-	await dataCacheStore.persist.rehydrate();
-	await locationStore.persist.rehydrate();
-
-	// 4. Sync contextStore state to VS Code context keys
-	contextStore.subscribe(({ state }) => {
-		vscode.commands.executeCommand("setContext", "isLoadingContextDirs", state.loading);
-		vscode.commands.executeCommand("setContext", "hasSelectedProject", !!state.selected);
-	});
-
-	// 5. Install and start the Choreo RPC server
-	await installRPCServer();
-
-	// 6. Create RPC client and wait for it to become active
-	const rpcClient = new ChoreoRPCClient();
-	await rpcClient.waitUntilActive();
-	ext.clients = { rpcClient };
-
-	// 7. Register authentication provider
-	const authProvider = new WSO2AuthenticationProvider(context.secrets, {
-		rpcClient,
-		log: ext.log.bind(ext),
-		logError: ext.logError.bind(ext),
-	});
-	ext.authProvider = authProvider;
-	context.subscriptions.push(
-		authentication.registerAuthenticationProvider(WSO2_AUTH_PROVIDER_ID, "WSO2 Integrator", authProvider, {
-			supportsMultipleAccounts: false,
-		}),
-	);
-
-	// 8. Sync auth state to VS Code context key
-	authProvider.subscribe(({ state }) => {
-		vscode.commands.executeCommand("setContext", "isLoggedIn", !!state.userInfo);
-	});
-
-	// 9. Initialize authentication (restores session from CLI if already signed in)
-	await authProvider.getState().initAuth();
-
-	// 10. Fetch remote config (billing/console URLs etc.)
-	ext.config = await ext.clients.rpcClient.getConfigFromCli();
-
-	// 11. Prompt restart when Advanced configuration keys change
-	registerPreInitHandlers();
-
-	// 12. Register VS Code commands
-	activateCmds(context);
-
-	// 13. Register URI handlers (sign-in callback, deep-link open)
-	activateURIHandlers();
-}
-
-function registerPreInitHandlers(): void {
-	workspace.onDidChangeConfiguration(async ({ affectsConfiguration }: ConfigurationChangeEvent) => {
-		if (
-			affectsConfiguration("WSO2.WSO2-Platform.Advanced.ChoreoEnvironment") ||
-			affectsConfiguration("WSO2.WSO2-Platform.Advanced.RpcPath")
-		) {
-			const selection = await window.showInformationMessage(
-				"WSO2 Platform extension configuration changed. Please restart VS Code for changes to take effect.",
-				"Restart Now",
-			);
-			if (selection === "Restart Now") {
-				if (affectsConfiguration("WSO2.WSO2-Platform.Advanced.ChoreoEnvironment")) {
-					ext.authProvider?.getState().logout();
-				}
-				commands.executeCommand("workbench.action.reloadWindow");
-			}
-		}
-	});
 }
 
 /**
